frequency to float, simplest with a keyboard with key input

+++ include/soloud_osc.h
@@ -42,7 +42,7 @@ namespace SoLoud
 		Osc();
 		~Osc();
 		result init(float aFreq);
-		result setFrequency(unsigned int aFreq);
+		result setFrequency(float aFreq);
 		float* getBlock(unsigned int size);
 
 	};
